Refactor: Rename 'features' folder to 'backend'
This renaming is made to prevent confusion.
